fix data structures and exporting
switch to using ids to represent relationships in json, since we now
have many to many relationships, so we can no longer have a hierarchical
structure.

also removed some files that won't be used immediately from the build.
TodayViewController is still not working, but it will be used so i left
it in the build.